intermediate timestamps and elapsed_time
Now timestamps and elapsed times work fine, and sorting is done
according to the furthest existing timestamp - which is refered to
as latest, but does not have to be.

Seeds are improved to reflect meaningful data, but all stamps are at
5, 10 and 15km, although the times are quarters, so actually for long
races this should have been 10,20 and 30km, but it works fine.

Most of time calculations are now done inside the run (which makes a
lot of sense) and some in the race (furthest/latest timestamp and
sorting).

Signed-off-by: Zeljko <zeljko@zwr.fi>