move all sorting and comparison to run
Having too many things in race is wrong, as those are features of
the run, so it is there now. Code looks cleaner.

In addition, as a very positive side effect, now runs are sorted by
all timestamps, from the last that the run has done, with nils being
last.

Signed-off-by: Zeljko <zeljko@zwr.fi>